feat(ocd): Refactor architecture for v1.0.0-rc
