+++ layouts/_partials/ocd/checks/frontmatter.logic.html
@@ -1,0 +1,31 @@
+{{- $errors := slice -}}
+{{- $context := . -}} 
+
+  {{- $filePath := "" -}}
+  {{- if .Page.File -}}
+    {{- $filePath = .Page.File.Path -}}
+  {{- end -}}
+
+  {{- $permalink := .Page.Permalink -}}
+
+{{- /* Check Page.Title */}}
+{{-  if not (strings.TrimSpace .Page.Title) }}
+  {{- $errorObject := dict
+    "code" "ocd-fm-100"
+    "page" $permalink
+    "file" $filePath 
+  -}}
+  {{- $errors = $errors | append $errorObject -}}
+{{- end -}}
+
+{{- /* Check Page.Params.description */ -}}
+{{- if eq .Page.Params.description "" -}}
+  {{- $errorObject := dict
+    "code" "ocd-fm-110"
+    "page" $permalink
+    "file" $filePath 
+  -}}
+  {{- $errors = $errors | append $errorObject -}}
+{{- end -}}
+
+{{- return $errors -}}
