+++ layouts/_partials/ocd/ui/console.html
@@ -54,7 +54,7 @@
     </div>
     <div class="ocd-console-content" data-alpha="ocd-console-content"></div>
     <footer>
-      {{- partial "ocd/components/local-versions.html" }}
+      {{- partial "ocd/components/version-display.html" }}
     </footer>
   </div>
 </div>
