+++ layouts/_shortcodes/hero.html
@@ -63,7 +63,7 @@
 -}}
 
 {{- /* *  Validate shortcode input */}}
-{{- $validationResult := partial "ocd/components/validator.html" (dict "rules" $rules "userInput" .Params) -}}
+{{- $validationResult := partial "ocd/core/validator.logic.html" (dict "rules" $rules "userInput" .Params) -}}
 
 {{- /* ** Conditions to render shortcode or OCD */}}
 {{- /* *** Input validation */}}
@@ -72,7 +72,7 @@
 {{- /* - On failed input validation call the OCD handler */}}
 {{- if not $validInput }}
   {{ if or ($ocdWarnEnabled) ($ocdEnabled) }}
-    {{- partial "ocd/components/handler.html" (dict
+    {{- partial "ocd/core/handler.logic.html" (dict
       "scName"   .Name
       "errors"   $validationResult.errors
       "rules"    $rules
@@ -122,7 +122,7 @@
         "link"            $link
         "linkTxt"         $linkTxt
       -}}
-      {{- partial "ocd/components/unique/link-pairs-sc.html" $errorData }}
+      {{- partial "ocd/views/errors/link-pairs.html" $errorData }}
     {{- end }}
   {{- else }}
