+++ layouts/_shortcodes/social.html
@@ -42,7 +42,7 @@
 -}}
 
 {{- /* *  Validate shortcode input */}}
-{{- $validationResult := partial "ocd/components/validator.html" (dict "rules" $rules "userInput" .Params) -}}
+{{- $validationResult := partial "ocd/core/validator.logic.html" (dict "rules" $rules "userInput" .Params) -}}
 
 {{- /* ** Conditions to render shortcode or OCD */}}
 {{- /* *** Input validation */}}
@@ -51,7 +51,7 @@
 {{- /* - On failed input validation call the OCD handler */}}
 {{- if not $validInput }}
   {{ if or ($ocdWarnEnabled) ($ocdEnabled) }}
-    {{- partial "ocd/components/handler.html" (dict
+    {{- partial "ocd/core/handler.logic.html" (dict
       "scName"   .Name
       "errors"   $validationResult.errors
       "rules"    $rules
@@ -78,7 +78,7 @@
         "menu"            "[[sc_social]]"
         "menuFile"        "config/_default/menus.toml"
       -}}
-      {{- partial "ocd/components/unique/missing-menu.html" $errorData }}
+      {{- partial "ocd/views/errors/missing-menu.html" $errorData }}
     {{- end }}
   {{- end }}
